Multiple changes:
- Shared memory cached node compaction (float4) with extended caching capacity.
- Fixed light leak problem of the plastic material (and add an switch for light leak, since light leak looks better for plant leaves)
- Fixed some BVH related bugs.

+++ src/renderer/depth.cuh
@@ -18,7 +18,7 @@ protected:
     int num_nodes;
     int num_cache;                  // number of cached BVH nodes
 
-    float4* _cached_nodes;
+    uint4* _cached_nodes;
     cudaTextureObject_t colormaps[3];
     cudaTextureObject_t bvh_leaves;
     cudaTextureObject_t nodes;
